move error handler to global domain
