+++ src/main/kotlin/com/example/webapp03/errorHandler/GlobalControlAdvice.kt
@@ -1,5 +1,6 @@
 package com.example.webapp03.errorHandler
 
+import org.springframework.dao.DuplicateKeyException
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Component
 import org.springframework.ui.Model
@@ -18,6 +19,17 @@ class GlobalControlAdvice {
         model.addAttribute("error", "Internal Server Error")
         model.addAttribute("message", "Happened Exception")
         model.addAttribute("status", HttpStatus.INTERNAL_SERVER_ERROR)
+
+        return "error"
+    }
+
+    @ExceptionHandler(DuplicateKeyException::class)
+    fun duplicateKeyExceptionHandler(
+        ex: DuplicateKeyException,
+        model: Model
+    ): String {
+        model.addAttribute("error", "Internal Server Error")
+        model.addAttribute("message", "DuplicateKeyException")
 
         return "error"
     }
